Support crazy HTTP servers that serve XML as HTML
Some HTTP servers are kind of crazy. They serve XML content with the
text/html Content-Type. Whether this is a bug in the server's code or
in its configuration doesn't help to get the feed content.

This change introduces the feed flag 'no-content-type'. When this flag
is set, feeds2email does not check for the Content-Type before parsing
the feed.