Avoid causing referral spam by using the Referer header
Feeds2email takes great care not to hammer servers with useless HTTP
requests using If-Modified-Since or If-None-Match headers. It also
reports itself in the User Agent and referer to publish its identity to
webmasters.

However, using the Referer HTTP header may cause spamming if that value
is printed in public logs or the like. So, this was a bad bad idea to
use it.

This change makes feeds2email uses only the User Agent header. Thanks to
Flameeyes for reporting the problem.

Fix isssue #1